AudioFX: add > 100% knob check
With how the knob logic is setup, while the knob is disabled, the user
can keep cranking it up, and it will save the progress. Upon enabling,
the knbo will try and restore that saved progress (which is > 100 now)
but since it was > 100 it will try and restore it to that same
value. Make sure it's less than 100 before enabling!

Change-Id: I40babd569c1c32dc20790820c3c2b8fe7ff88a18
Signed-off-by: Roman Birg <roman@cyngn.com>