Update patch set 6
Patch Set 6:

Steve: The "goes up to eleven" reference on the original is nice. It could be adjusted a bit to fit a little better though. Basically by making it look slightly less like a real button and by making the shadows more material-like. That should not be too hard if the original designed kept all the source material and not just the finished product.

Patch-set: 6