AudioFX: fix automatic output device switching
This patch fixes when leaving the activity, switching devices 
(e.g. plugging headphones in/out), and then returning the activity, 
the output device is not updated.

Whenever the Activity resumes, it tries to bind to the service, but if 
this service was already bound, its state might have changed, and we 
are not reading the new state from the service. With this patch, every 
call to bindService() will ensure that the client side is in the 
correct state.

Change-Id: Ia362759c03f54037769555bed7ea2db2f0d43515
Signed-off-by: Roman Birg <roman@cyngn.com>