AudioFX: use a handler to handle session work
Use a handler that has a priority equal to/higher than the audio thread
to ensure the effects are set before music starts playing.

Without this patch, on some devices, lowering the volume, pausing, and
starting a song would cause a loud pop to happen, as the effects were
likely destroyed and not yet applied to the session before it started
playing.

Change-Id: I729bdc270367edf099ddc840eda37cefd2b7a74a
Signed-off-by: Roman Birg <roman@cyngn.com>