AudioFX: cleanup logic handling persistent service
Since we no longer have a persistent service, the app can be killed as
necessary, so we no longer have to do any kind of detection to see
whether the defaults were initiated yet. When the data is cleared, the
UI will be torn down and back up as needed.

Also we no longer need the initial animation, as the UI will have all
its initial defaults ready to go as soon as the fragments are added.

Change-Id: Iaf3b9d5b2f5ad9bcd2a52fb84274743fa40cadf7
Signed-off-by: Roman Birg <roman@cyngn.com>