AudioFX: persist user device selection across rotates
We want to allow users to play with other devices while they may not
necessarily be active. We should also persist the selection across
rotates. Currently every time the service is bound, the current device
was read from the service - set to the currently _active_ device.

Now we have two separate devices we keep track of: the real output, and
the user selection. Keep the user selection around for the duration of
the activity lifecycle, only persisting it in save instance state.

Change-Id: If631d56d901c2a11bb9e06c33122fd3d90d49a77
Signed-off-by: Roman Birg <roman@cyngn.com>