AudioFX: fix current bluetooth device logic
Since the service is no longer persisted, this bug became painfully
obvious. When more than one device is connected to bluetooth, it doesn't
properly select the connected one, but the one that was last changed.

Instead, we should check for one bonded AND connected device, and always
return that one.

Change-Id: I0e3d909c4287a3b456c3d278353b0b4611c2db1a
Signed-off-by: Roman Birg <roman@cyngn.com>