AudioFX: update audio service eq bands as they move
With this commit, switching between EQ bands should sound as smooth as
the UI change looks.

We need to use a background work handler in the service to make sure
that the work does NOT affect the UI, which will probably be doing a bit
of work when this happens.

Change-Id: I12a5acee3a90556e1244d8dd382f743e4cf4feec
Signed-off-by: Roman Birg <roman@cyngn.com>