Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

AudioFX shouldn't need runtime permissions, they should be granted to it by default. Since it's already marked as a persistent app in AndroidManifest.xml all that it really needs is to be platform signed to get the permissions it wants. Add this to Android.mk and do a new clean build:

LOCAL_CERTIFICATE := platform

Patch-set: 6