Update patch set 7
Patch Set 7:

What's the problem with it being platform signed? This app is for CyanogenMod and not meant to be installed or upgraded as a 3rd party app anywhere else.


If you look at that DefaultPermissionGrantPolicy, most gapps fall under one of those categories listed. Some do not and request permissions (e.g. Hangouts).


Browser and sound recorder seem very different from AudioFX to me - they are not persistent and they don't need to be running in the background. 

AudioFX may be on by default, and shouldn't need users' permission to enable sound effects in the background.

Patch-set: 7