AudioFX: fix clearing app data
After the recent service changes, clearing app data would cause two
equalizer fragments to be on top of each other.

Only listen to the preference we want as we may get callbacks when all
values are set (so defaultsSet would be true for both keys if checked),
so it would apply init() twice.

Also cleaned up some initialization to catch the new defaults.

Ref: AUDIO-142

Change-Id: I2286043e910dca786fecaa5775aba770d877a420
Signed-off-by: Roman Birg <roman@cyngn.com>