AudioFX: fix not setting up proper defaults on fresh boot
The service will setup its own defaults as needed. If there is a
necessary udpate to prefs, the pref change increment logic will handle
that and we don't explicitly need to call to reapply defaults on every
boot. The service will do it in every onCreate() AS NECESSRY!

This logic was also leftover from before there was a preference version
code.

Ref: AUDIO-145, OPO-512

Change-Id: I950bed1fbb6289cd2bcc53bac8ee0c268e225774
Signed-off-by: Roman Birg <roman@cyngn.com>