AudioFX: fix missing bassboost/virtualizer
If the device doesn't have headphones plugged in, and the speaker
reports that it doesn't support virtualizer strength in speaker mode, we
made the assumption it was never available. Just check for a valid
refernce to the effect, all AndroidEffects _should_ have a virtualizer.

Also moved the prefs version to the prefs manager.

Change-Id: Id44087c509ece3d55b42c985660a61222bbdfe17
Ticket: AUDIO-152
Signed-off-by: Roman Birg <roman@cyngn.com>