Update patch set 2
Patch Set 2:

I think it does permit long-clicks. Looking at the code in SystemUI, it's trying to resolve an application-level intent for TileService.ACTION_QS_TILE_PREFERENCES. If it can't find this, it uses the appinfo screen. It doesn't seem to be able to deal with an arbitrary user-supplied component though.

I don't see any way to do what we currently do with AudioFX- publish a tile automatically not based on a user action. For AudioFX we do it when there's an active audio session. I don't think it's so great for any app to be able to do this because of the potential for abuse, but it's needed for trusted system apps. 

I'm not sure if we used the RemoteViews for anything before, I vaguely remembering it being part of a hackathon project though.

Still going thru the new code before I take a swing at refactoring the SDK.

Patch-set: 2
Reviewer: Steve Kondik <32@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0