Update patch set 3
Patch Set 3: Code-Review+1

For a lot of the streams no flags are set (flags=FLAG_NONE == 0). No flags simply means "select the primary output". This is what happens with many streams of type STREAM_MUSIC (e.g. YouTube). The actual flags will be set later on, but the AudioSession API will never report these changes and in any case I don't think that we should rely on the output flags to decide whether or not we should apply effects.

Patch-set: 3
Label: Code-Review=+1